refactor: uploaded file status
