+++ web-ui/src/api/models/uploaded-file.js
@@ -1,22 +1,6 @@
 import Cytomine from '../cytomine.js';
 import Model from './model.js';
 
-/** Enum providing the uploaded file status handled in Cytomine */
-export const UploadedFileStatus = Object.freeze({
-  UPLOADED: 0,
-  DETECTING_FORMAT: 10,
-  ERROR_FORMAT: 11,
-  EXTRACTING_DATA: 20,
-  ERROR_EXTRACTION: 21,
-  CONVERTING: 30,
-  ERROR_CONVERSION: 31,
-  DEPLOYING: 40,
-  ERROR_DEPLOYMENT: 41,
-  DEPLOYED: 100,
-  EXTRACTED: 102,
-  CONVERTED: 104
-});
-
 export default class UploadedFile extends Model {
   /** @inheritdoc */
   static get callbackIdentifier() {
