+++ web-ui/src/constants/UploadedFileStatus.js
@@ -1,0 +1,19 @@
+export const UploadedFileStatus = {
+  UPLOADED: 0,
+  DETECTING_FORMAT: 10,
+  ERROR_FORMAT: 11,
+  EXTRACTING_DATA: 20,
+  ERROR_EXTRACTION: 21,
+  CONVERTING: 30,
+  ERROR_CONVERSION: 31,
+  DEPLOYING: 40,
+  ERROR_DEPLOYMENT: 41,
+  UNPACKING: 50,
+  ERROR_UNPACKING: 51,
+  CHECKING_INTEGRITY: 60,
+  ERROR_INTEGRITY: 61,
+  DEPLOYED: 100,
+  EXTRACTED: 102,
+  CONVERTED: 104,
+  UNPACKED: 106,
+};
